+++ src/Data/Profunctor/Tambara.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE CPP #-
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE Rank2Types #-}
 {-# LANGUAGE TypeOperators #-}